requirements.txt
# Core dependencies for experiments
jupyter>=1.0.0
notebook>=6.0.0
numpy>=1.19.0
matplotlib>=3.3.0
pandas>=2.1.3
seaborn>=0.11.0
h5py>=3.10.0
python-dotenv>=1.0.0
tqdm>=4.67.1

# RWalks package - install from local directory
./rwalks/
